backend/requirements.txt
setuptools>=68.0.0
wheel>=0.40.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.0.2
Pillow>=10.0.0
httpx>=0.24.0,<0.25.0
openai==1.3.7
requests==2.31.0
numpy>=1.26.0
pydantic>=2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
amadeus==8.1.0
gradio-client==1.10.4
